chore: Update RareLink FHIR validation data
This commit updates the FHIR validation data for the RareLink project. It adds new observation and condition data to the validation set. The changes ensure that the validation data is up to date and reflects the latest requirements of the project.

Note: You need Python ≥3.10 but <3.13 to be able to use RareLink!