160 ig publisher (#166)
* adding first files

* added input fsh

* added  temp

* added other files

* output folder

* added

* added

* add

* fix

* update to v2.0.0

* chore: Update version to 2.0.0

* updated instances

* chore: Update virtual environment setup script

* chore: Update RareLink FHIR validation data

This commit updates the FHIR validation data for the RareLink project. It adds new observation and condition data to the validation set. The changes ensure that the validation data is up to date and reflects the latest requirements of the project.

Note: You need Python ≥3.10 but <3.13 to be able to use RareLink!

* evaluation cohort phenopackets